Added required type casting.

+++ inc/benchmark.php
@@ -26,7 +26,7 @@ function get_human_readable_usage($size) {
     if ($size < 1024)
         return $result;
 
-    $index = floor(log($size, 1024));
+    $index = (int)floor(log($size, 1024));
     if ($index < sizeof($units)) {
         $result = round($size / pow(1024, ($index)), 2) . ' ' . $units[$index];
     }
